# docs/swagger.yaml
definitions:
...
  MenuPayloads.WeightHistoryPayloads:
    properties:
      user_bmi:
        type: number
      user_weight:
        description: UserId         int       `gorm:"column:user_id" json:"user_id"`
        type: number
      user_weight_time:
        type: string
    type: object
...
  entities.UserDetail:
    properties:
      user_detail_id:
        type: integer
      user_gender:
        type: string
      user_height:
        type: number
      user_id:
        type: integer
      user_phone_number:
        description: UserProfileImage       string  `gorm:"column:user_profile_image"
          json:"user_profile_image"`
        type: string
      user_profile_description:
        type: string
      user_weight:
        type: number
    type: object
  entities.Users:
    properties:
      user_detail:
        allOf:
        - $ref: '#/definitions/entities.UserDetail'
        description: IsVIP        bool       `gorm:"column:is_vip" json:"is_vip"`
      user_email:
        type: string
      user_id:
        type: integer
      user_name:
        type: string
      user_password:
        type: string
    type: object
  entities.WeightHistoryEntities:
    properties:
      user:
        $ref: '#/definitions/entities.Users'
      user_bmi:
        type: number
      user_id:
        type: integer
      user_weight:
        type: number
      user_weight_time:
        type: string
      weight_history_id:
        type: integer
    type: object
...
  responses.ErrorResponses:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
  responses.StandarAPIResponses:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
      success:
        type: boolean
    type: object
host: localhost:3000
info:
  contact:
    email: devin@gmail.com
    name: IoT Blynk Smart Agriculture Backend Service
    url: IoT
  description: IoT Blynk Smart Agriculture Backend Service
  license:
    name: IoT Blynk Smart Agriculture Backend Service
  termsOfService: http://swagger.io/terms/
  title: IoT Blynk Smart Agriculture Backend Service
  version: "1.0"
paths:
...
  /api/weight:
    get:
      consumes:
      - application/json
      description: Get Weight Pagination History
      parameters:
      - description: user_id
        in: path
        name: user_id
        required: true
        type: integer
      - description: sort_by
        in: query
        name: sort_by
        type: string
      - description: sort_of
        in: query
        name: sort_of
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.WeightHistoryEntities'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
      security:
      - BearerAuth: []
      summary: Get Weight Pagination History
      tags:
      - Weight
    post:
      consumes:
      - application/json
      description: Create Weight History
      parameters:
      - description: Insert Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/MenuPayloads.WeightHistoryPayloads'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.StandarAPIResponses'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponses'
      security:
      - BearerAuth: []
      summary: Create Weight History
      tags:
      - Weight
  /api/weight/delete/{weight_id}:
    delete:
      consumes:
      - application/json
      description: Create Weight History
      parameters:
      - description: weight_id
        in: path
        name: weight_id
        required: true
        type: integer
      - description: user_id
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.StandarAPIResponses'
      security:
      - BearerAuth: []
      summary: Create Weight History
      tags:
      - Weight
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"